refactor(common): 抽取 atomic_write 到 common::fs 供 model 层复用
atomic_write（写临时文件 → fsync → rename 替换 → fsync 父目录）此前是
token_manager.rs 的模块私有函数。后续 API Key ID 计数器同样需要原子写，
但 model 层不应反向依赖 kiro 层，故整段原样移入 src/common/fs.rs，
签名改为 pub(crate)。

行为不变：token_manager.rs 的 3 处调用点均为非限定调用，仅新增一行 use。
