+++ src/common/mod.rs
@@ -2,3 +2,4 @@
 //! 公共工具模块
 
 pub mod auth;
+pub mod fs;
