+++ src/kiro/token_manager.rs
@@ -16,6 +16,7 @@ use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
 use std::time::{Duration as StdDuration, Instant};
 
+use crate::common::fs::atomic_write;
 use crate::http_client::{ProxyConfig, build_client};
 use crate::kiro::machine_id;
 use crate::kiro::model::available_models::AvailableModelsResponse;
@@ -876,33 +877,6 @@ struct StickyCacheEntry {
     /// 仅当连续限流达到 STICKY_THROTTLE_EVICT_THRESHOLD 才判定该账号确实不适合
     /// 承载此会话，执行解绑重选。任一次成功即清零。
     consecutive_throttles: u32,
-}
-
-/// 原子写文件：写临时文件 → fsync → rename 替换 → fsync 父目录。
-/// 同目录 rename 在 POSIX 上是原子操作，避免写半截导致目标文件损坏。
-fn atomic_write(path: &std::path::Path, bytes: &[u8]) -> std::io::Result<()> {
-    use std::io::Write;
-    let dir = path.parent().unwrap_or_else(|| std::path::Path::new("."));
-    let tmp = path.with_extension(format!("tmp.{}", std::process::id()));
-
-    // 写临时文件并落盘
-    {
-        let mut file = std::fs::File::create(&tmp)?;
-        file.write_all(bytes)?;
-        file.sync_all()?;
-    }
-
-    // 原子替换目标文件
-    if let Err(e) = std::fs::rename(&tmp, path) {
-        let _ = std::fs::remove_file(&tmp); // 清理残留临时文件
-        return Err(e);
-    }
-
-    // fsync 父目录，确保 rename 元数据落盘（容器持久化卷必须）
-    if let Ok(dir_file) = std::fs::File::open(dir) {
-        let _ = dir_file.sync_all();
-    }
-    Ok(())
 }
 
 /// API 调用上下文
